+++ compat/strlcpy.c
@@ -47,5 +47,6 @@ strlcpy(char *dst, const char *src, size_t size)
 		while (*src++);
 	}
 
-	return src - s - 1;
+
+	return (size_t)(src - s - 1);
 }
